requirements.txt
beautifulsoup4>=4.9.0
requests>=2.24.0
pyyaml>=5.3.1